refactor(node-models): Reorganized shared node models
