--- src/app/features/my-projects/models/create-project.model.ts
@@ -1,25 +1,0 @@
-export interface CreateProjectPayloadJsoApi {
-  data: {
-    type: 'nodes';
-    attributes: {
-      title: string;
-      description?: string;
-      category: 'project';
-      template_from?: string;
-    };
-    relationships: {
-      region: {
-        data: {
-          type: 'regions';
-          id: string;
-        };
-      };
-      affiliated_institutions?: {
-        data: {
-          type: 'institutions';
-          id: string;
-        }[];
-      };
-    };
-  };
-}
