+++ src/app/features/my-projects/models/index.ts
@@ -1,4 +1,3 @@
-export * from './create-project.model';
 export * from './my-projects.models';
 export * from './my-projects-endpoint.type';
 export * from './my-projects-search-filters.models';
