+++ src/app/features/my-projects/services/my-projects.service.ts
@@ -7,11 +7,10 @@ import { JsonApiResponse } from '@core/models';
 import { JsonApiService } from '@osf/core/services';
 import { SparseCollectionsResponseJsonApi } from '@osf/features/collections/models';
 import { ResourceType, SortOrder } from '@osf/shared/enums';
-import { NodeResponseModel, UpdateNodeRequestModel } from '@shared/models';
+import { CreateProjectPayloadJsoApi, NodeResponseModel, UpdateNodeRequestModel } from '@shared/models';
 
 import { MyProjectsMapper } from '../mappers';
 import {
-  CreateProjectPayloadJsoApi,
   EndpointType,
   MyProjectsItem,
   MyProjectsItemGetResponseJsonApi,
