+++ src/app/shared/models/index.ts
@@ -4,7 +4,6 @@ export * from './brand.model';
 export * from './charts';
 export * from './confirmation-options.model';
 export * from './create-component-form.model';
-export * from './create-project-form.model';
 export * from './file-menu-action.model';
 export * from './files/file.model';
 export * from './files/get-files-response.model';
@@ -19,8 +18,9 @@ export * from './license.model';
 export * from './licenses-json-api.model';
 export * from './metadata-field.model';
 export * from './nav-item.model';
-export * from './node-response.model';
 export * from './node-subject.model';
+export * from './nodes/create-project-form.model';
+export * from './nodes/nodes-json-api.model';
 export * from './paginated-data.model';
 export * from './query-params.model';
 export * from './resource-card';
@@ -37,6 +37,5 @@ export * from './table-parameters.model';
 export * from './toolbar-resource.model';
 export * from './tooltip-position.model';
 export * from './tutorial-step.model';
-export * from './update-node-request.model';
 export * from './user';
 export * from './validation-params.model';
