--- src/app/shared/models/update-node-request.model.ts
@@ -1,16 +1,0 @@
-export interface UpdateNodeAttributes {
-  description?: string;
-  tags?: string[];
-  public?: boolean;
-  title?: string;
-}
-
-export interface UpdateNodeData {
-  type: 'nodes';
-  id: string;
-  attributes: UpdateNodeAttributes;
-}
-
-export interface UpdateNodeRequestModel {
-  data: UpdateNodeData;
-}
